refactored themeHandler into action
