+++ action/popup/popup.html
@@ -8,7 +8,7 @@
 		<link rel="stylesheet" href="../variables.css" />
 		<link rel="stylesheet" href="./popup.css" />
 		<link rel="stylesheet" href="../basic.css" />
-		<script src="../../themeHandler.js" type="module"></script>
+		<script src="../themeHandler.js" type="module"></script>
 		<script src="./popup.js" type="module"></script>
 		<script src="../../dragHandler.js" type="module"></script>
 	</head>
